user/requirements.txt
anyio==3.6.2
astroid==2.3.3
autopep8==1.5.5
bcrypt==4.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.3
decorator==4.4.2
distlib==0.3.6
dnspython==2.2.1
fastapi==0.85.2
filelock==3.8.0
Flask==2.0.2
future==0.18.2
h11==0.14.0
idna==3.3
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
networkx==2.5
numpy==1.20.1
opencv-python==4.5.5.64
pandas==1.2.3
passlib==1.7.4
Pillow==8.2.0
platformdirs==2.5.2
psutil==5.9.3
pycodestyle==2.6.0
pydantic==1.10.2
pyglet==1.4.10
PyJWT==2.6.0
pylint==2.4.4
pymongo==4.3.2
python-dateutil==2.8.1
python-decouple==3.6
pytz==2021.1
requests==2.26.0
six==1.13.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.7
uvicorn==0.19.0
virtualenv==20.16.6
Werkzeug==2.0.2
wrapt==1.11.2
